feat: split main.go into more files, update makefile to support build and run separately
