+++ server/go.mod
@@ -1,9 +1,8 @@
-module src/main.go
+module openorganizer
 
 go 1.24.2
 
 require (
-	github.com/jmoiron/sqlx v1.4.0
 	github.com/joho/godotenv v1.5.1
 	github.com/lib/pq v1.10.9
 )
